# docs/swagger.yaml
basePath: /v1
definitions:
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.LeadDto:
    properties:
      comment:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      doctorId:
        type: integer
      email:
        description: E-mail string.
        example: mail@mail.com
        type: string
      name:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      phone:
        description: E.164 phone number string.
        example: "+79876543210"
        type: string
      status:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      type:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
    type: object
...
  dto.PetDto:
    properties:
      breed:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      clientId:
        type: integer
      colour:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
      name:
        description: Alphabetic characters only (Russian and English).
        maxLength: 255
        type: string
      sex:
        description: Alphabetic characters only (Russian and English).
        type: string
      type:
        description: 'Allowed characters: printable ASCII (Russian and English).'
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Client:
    properties:
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      sex:
        type: string
      surname:
        type: string
      updated_at:
        type: string
    type: object
  models.Department:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      slug:
        type: string
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Lead:
    properties:
      comment:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      doctor:
        $ref: '#/definitions/models.User'
      doctorId:
        type: integer
      email:
        type: string
      id:
        type: integer
      lastUpdatedBy:
        $ref: '#/definitions/models.User'
      lastUpdatedById:
        type: integer
      name:
        type: string
      phone:
        type: string
      status:
        description: open, in_progress, closed, rejected
        type: string
      type:
        description: in clinic, online, callback
        type: string
      updated_at:
        type: string
    type: object
  models.Pet:
    properties:
      breed:
        type: string
      client:
        $ref: '#/definitions/models.Client'
      clientId:
        type: integer
      colour:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      sex:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Service:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      active:
        type: boolean
      birthDate:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      departments:
        items:
          $ref: '#/definitions/models.Department'
        type: array
      email:
        type: string
      id:
        type: integer
      info:
        type: string
      name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      profession:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      sex:
        type: string
      slug:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vic.pozdeev@gmail.com
    name: Victor Pozdeev
  description: This is API specification for vet-clinic API server.
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.php
  title: Vet clinic API
  version: v0.1.0
paths:
...
  /leads:
    get:
      consumes:
      - application/json
      description: Returns the list of leads.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            items:
              $ref: '#/definitions/models.Lead'
            type: array
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a lead list.
      tags:
      - Leads
    post:
      consumes:
      - application/json
      description: Create a new lead.
      parameters:
      - description: A new lead data for creating.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LeadDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Lead'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new lead.
      tags:
      - Leads
  /leads/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the existing lead.
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Lead'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Delete the existing lead. Required user''s role: Superuser'
      tags:
      - Leads
    get:
      consumes:
      - application/json
      description: Returns one record matched lead's id.
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Lead'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a lead.
      tags:
      - Leads
    put:
      consumes:
      - application/json
      description: Update the existing lead.
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      - description: Lead data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LeadDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Lead'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Update the existing lead.
      tags:
      - Leads
...
  /pets:
    get:
      consumes:
      - application/json
      description: Returns the list of pets.
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            items:
              $ref: '#/definitions/models.Pet'
            type: array
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a pet list.
      tags:
      - Pets
    post:
      consumes:
      - application/json
      description: Create a new pet.
      parameters:
      - description: A new pet data for creating.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.PetDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Create a new pet.
      tags:
      - Pets
  /pets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the existing pet.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
        "403":
          description: Access denied.
      security:
      - ApiKeyAuth: []
      summary: 'Delete the existing pet. Required user''s role: Superuser'
      tags:
      - Pets
    get:
      consumes:
      - application/json
      description: Returns one record matched pet's id.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Failed to fetch data.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Get a pet.
      tags:
      - Pets
    put:
      consumes:
      - application/json
      description: Update the existing pet.
      parameters:
      - description: Pet ID
        in: path
        name: id
        required: true
        type: string
      - description: Pet data for update.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.PetDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success to fetch data.
          schema:
            $ref: '#/definitions/models.Pet'
        "400":
          description: Failed to the registration.
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Failed to the authentication.
      security:
      - ApiKeyAuth: []
      summary: Update the existing pet.
      tags:
      - Pets
...
schemes:
- http
swagger: "2.0"